requirements.txt
pyserial
pyqt5
qt5-tools